backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
passlib==1.7.4
sentence-transformers==3.1.1
google-generativeai==0.8.3
numpy==1.26.4
scikit-learn==1.5.2
yfinance==0.2.44